+++ packages/core-js-compat/src/data.mjs
@@ -2635,7 +2635,9 @@ export const data = {
     safari: '14.0',
   },
   'web.url.can-parse': {
-    bun: '1.0.2',
+    // Bun ~ 1.0.30 bug
+    // https://github.com/oven-sh/bun/issues/9250
+    // bun: '1.0.2',
     chrome: '120',
     deno: '1.33.2',
     firefox: '115',